Seperate out the low level api examples and higher level dsl examples
