+++ examples/src/java/BUILD
@@ -1,8 +1,8 @@
 package(default_visibility = ["//visibility:public"])
 
 java_binary(
-    name='examples-unshaded',
-    srcs = glob(["**/*.java"]),
+    name='api-examples-unshaded',
+    srcs = glob(["com/twitter/heron/examples/api/**/*.java"]),
     deps = [
         "//heron/api/src/java:api-java",
         "//heron/common/src/java:basics-java",
@@ -12,8 +12,26 @@ java_binary(
 )
 
 genrule(
-    name = 'heron-examples',
-    srcs = [":examples-unshaded_deploy.jar"],
-    outs = ["heron-examples.jar"],
+    name = 'heron-api-examples',
+    srcs = [":api-examples-unshaded_deploy.jar"],
+    outs = ["heron-api-examples.jar"],
+    cmd  = "cp $< $@",
+)
+
+java_binary(
+    name='dsl-examples-unshaded',
+    srcs = glob(["com/twitter/heron/examples/dsl/**/*.java"]),
+    deps = [
+        "//heron/api/src/java:api-java",
+        "//heron/common/src/java:basics-java",
+        "//heron/simulator/src/java:simulator-java"
+    ],
+    create_executable = 0,
+)
+
+genrule(
+    name = 'heron-dsl-examples',
+    srcs = [":dsl-examples-unshaded_deploy.jar"],
+    outs = ["heron-dsl-examples.jar"],
     cmd  = "cp $< $@",
 )
