+++ examples/src/java/com/twitter/heron/examples/api/ExampleResources.java
@@ -11,7 +11,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package com.twitter.heron.examples;
+package com.twitter.heron.examples.api;
 
 import com.twitter.heron.common.basics.ByteAmount;
