+++ examples/src/java/com/twitter/heron/examples/api/bolt/SlidingWindowSumBolt.java
@@ -11,7 +11,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package com.twitter.heron.examples.bolt;
+package com.twitter.heron.examples.api.bolt;
 
 import java.util.List;
 import java.util.Map;
