+++ examples/src/java/com/twitter/heron/examples/api/spout/RandomIntegerSpout.java
@@ -11,7 +11,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package com.twitter.heron.examples.spout;
+package com.twitter.heron.examples.api.spout;
 
 import java.util.Map;
 import java.util.Random;
