+++ examples/src/java/com/twitter/heron/examples/api/spout/TestWordSpout.java
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package com.twitter.heron.examples.spout;
+package com.twitter.heron.examples.api.spout;
 
 import java.time.Duration;
 import java.util.Map;
